refactorización del código del dado
